+++ src/vibe3/services/flow/block_mixin.py
@@ -36,7 +36,7 @@ class FlowLifecycleMixin:
             repo: Repository (defaults to current repo)
             event_type: Event type for timeline ("flow_blocked" or "flow_failed")
         """
-        from vibe3.services.blocked_state_service import BlockedStateService
+        from vibe3.services.flow.blocked_state_service import BlockedStateService
 
         logger.bind(
             domain="flow",
